src/mfi_ddb/retrieval_api/rws/requirements.txt
FastAPI[all]
psycopg2-binary